Bug fix in phone integer size

+++ delivery/models.py
@@ -8,7 +8,7 @@ class Clients(models.Model):
     street_name = models.CharField(max_length=200, verbose_name='Calle')
     street_number = models.CharField(max_length=200, verbose_name='Número')
     street_extra = models.CharField(max_length=200, blank=True, verbose_name='Indicaciones')
-    phone = models.PositiveIntegerField(verbose_name='Teléfono')
+    phone = models.PositiveBigIntegerField(verbose_name='Teléfono')
     email = models.EmailField(blank=True, verbose_name='Email')
     days_between_visits = models.PositiveIntegerField(default=7, verbose_name='Días entre visitas')
     registered = models.DateField(default=datetime.date.today, verbose_name='Fecha de alta')
